Extract MqttPacketCreator from PlainMessageReader and SslMessageReader
