+++ core-service/src/main/java/javasabr/mqtt/service/impl/PlainMqttConnectionFactory.java
@@ -4,6 +4,7 @@ import java.nio.channels.AsynchronousSocketChannel;
 import javasabr.mqtt.model.MqttServerConnectionConfig;
 import javasabr.mqtt.network.MqttConnection;
 import javasabr.mqtt.network.MqttConnectionFactory;
+import javasabr.mqtt.network.message.MqttPacketCreator;
 import javasabr.mqtt.network.user.NetworkMqttUserFactory;
 import javasabr.rlib.network.Network;
 import javasabr.rlib.network.impl.DefaultBufferAllocator;
@@ -13,11 +14,12 @@ import lombok.experimental.FieldDefaults;
 
 @RequiredArgsConstructor
 @FieldDefaults(level = AccessLevel.PROTECTED, makeFinal = true)
-public class DefaultMqttConnectionFactory implements MqttConnectionFactory<MqttConnection> {
+public class PlainMqttConnectionFactory implements MqttConnectionFactory<MqttConnection> {
 
   MqttServerConnectionConfig serverConnectionConfig;
   NetworkMqttUserFactory clientFactory;
   int maxPacketsByRead;
+  MqttPacketCreator mqttPacketCreator;
 
   @Override
   public MqttConnection newConnection(Network<MqttConnection> network, AsynchronousSocketChannel channel) {
@@ -28,6 +30,7 @@ public class DefaultMqttConnectionFactory implements MqttConnectionFactory<MqttC
         bufferAllocator,
         maxPacketsByRead,
         serverConnectionConfig,
-        clientFactory);
+        clientFactory,
+        mqttPacketCreator);
   }
 }
