+++ core-service/src/test/groovy/javasabr/mqtt/service/IntegrationServiceSpecification.groovy
@@ -10,6 +10,7 @@ import javasabr.mqtt.model.QoS
 import javasabr.mqtt.model.publish.PublishData
 import javasabr.mqtt.network.MqttConnection
 import javasabr.mqtt.network.handler.NetworkMqttUserReleaseHandler
+import javasabr.mqtt.network.message.MqttPacketCreator
 import javasabr.mqtt.service.impl.DefaultMessageOutFactoryService
 import javasabr.mqtt.service.impl.DefaultTopicService
 import javasabr.mqtt.service.impl.DisabledAuthorizationService
@@ -171,7 +172,8 @@ abstract class IntegrationServiceSpecification extends BaseSpecification {
           user.session(createdSession)
           user.clientId(generatedClientId)
           return user
-        })
+        },
+        new MqttPacketCreator())
 
     connection.configure(new MqttClientConnectionConfig(
         serverConnectionConfig,
@@ -201,6 +203,7 @@ abstract class IntegrationServiceSpecification extends BaseSpecification {
         serverConnectionConfig,
         { MqttConnection ownedConnection ->
           return new TestExternalNetworkMqttUser(ownedConnection, Mock(NetworkMqttUserReleaseHandler))
-        })
+        },
+        new MqttPacketCreator())
   }
 }
