+++ network/src/main/java/javasabr/mqtt/network/message/ssl/SslMqttMessageWriter.java
@@ -1,4 +1,4 @@
-package javasabr.mqtt.network.ssl;
+package javasabr.mqtt.network.message.ssl;
 
 import java.nio.ByteBuffer;
 import java.util.function.Consumer;
